// models/Broker/Quotes/QuotesDetails.ts
export interface QuotesDetails {
    id: number;
    cargo_weight: string;
    destination: string;
    load_id: number;
    broker_id: number;
    cargo_type: string;
    status: string;
    equipment_type: string;
    pickup_time: string;
    delivery_time: string;
    new_quote: number;
    quote: string;
    origin: string;
    updated_date: string;
    name: string;
    last_quote: number;
    customer_phone_number: string;
    customer_email: string;
}